feat: migrate project to Next.js
