+++ package.json
@@ -26,8 +26,7 @@
     "react-icons": "^3.10.0",
     "react-loading-skeleton": "^2.1.1",
     "react-lottie": "^1.2.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
+    "next": "13.4.19",
     "react-spring": "^8.0.27",
     "react-switch": "^5.0.1",
     "react-transition-group": "^4.4.1",
@@ -36,10 +35,10 @@
     "uuidv4": "^6.1.1"
   },
   "scripts": {
-  "start": "react-scripts --openssl-legacy-provider start",
-  "build": "react-scripts --openssl-legacy-provider build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "test": "echo \"No tests specified\" && exit 0"
   },
   "browserslist": {
     "production": [
@@ -55,7 +54,6 @@
   },
   "devDependencies": {
     "@types/react-lottie": "^1.2.5",
-    "@types/react-router-dom": "^5.1.5",
     "@types/react-transition-group": "^4.4.1",
     "@types/styled-components": "^5.1.1",
     "@typescript-eslint/eslint-plugin": "^2.30.0",
