--- src/App.tsx
@@ -1,33 +1,0 @@
-import React,{useEffect} from 'react';
-import { BrowserRouter } from 'react-router-dom';
-import { ThemeProvider, DefaultTheme } from 'styled-components';
-import usePeristedState from './utils/usePersistedState';
-import Routes from './routes';
-import GlobalStyle from './styles/global';
-import light from './styles/themes/light';
-import dark from './styles/themes/dark';
-import analytics from '@vercel/analytics';
-
-const App: React.FC = () => {
-  const [theme, setTheme] = usePeristedState<DefaultTheme>('theme', light);
-
-  const toggleTheme = () => {
-    setTheme(theme.title === 'light' ? dark : light);
-  };
-
-  useEffect(() => {
-    analytics.track('Home');
-  }, []);
-
-
-  return (
-    <ThemeProvider theme={theme}>
-      <BrowserRouter>
-        <Routes toggleTheme={toggleTheme} />
-        <GlobalStyle />
-      </BrowserRouter>
-    </ThemeProvider>
-  );
-};
-
-export default App;
