+++ src/pages/_app.tsx
@@ -1,0 +1,21 @@
+import type { AppProps } from 'next/app';
+import { ThemeProvider, DefaultTheme } from 'styled-components';
+import { useState } from 'react';
+import GlobalStyle from '../styles/global';
+import light from '../styles/themes/light';
+import dark from '../styles/themes/dark';
+
+export default function MyApp({ Component, pageProps }: AppProps) {
+  const [theme, setTheme] = useState<DefaultTheme>(light);
+
+  const toggleTheme = () => {
+    setTheme(theme.title === 'light' ? dark : light);
+  };
+
+  return (
+    <ThemeProvider theme={theme}>
+      <Component {...pageProps} toggleTheme={toggleTheme} />
+      <GlobalStyle />
+    </ThemeProvider>
+  );
+}
